Update to 14.0.01 release of GNU Unifont.
I actually deleted the GitHub repository in preparation to release an alternative approach that didn't embed the font data in the executable. I was going to do this out of concern over the possibility of enabling license violations. As stated in the readme, I believe the crate's current usage of GNU Unifont is not in violation of the font's licensing terms, but I was (and am) not totally certain about this. This uncertainty was enough to prompt investigating an alternative approach. I quickly realized that the alternative would slightly increase overhead and/or significantly complicate the lifetime management aspect of using this crate, which strongly outweighed my worries.

Again, if any of the copyright holders object to this crate's usage of GNU Unifont, please contact me and we can discuss it.